deleting user news from admin

+++ socialNetwork/ui/src/main/java/by/academy/it/controllers/controller/UserInfoController.java
@@ -42,8 +42,8 @@ public class UserInfoController {
         String status = friendsService.getFriendStatus(authorizedUser.getUserId(), friendId);
 
         ModelAndView modelAndView = new ModelAndView("user-info")
-                .addObject("subscribersCount",userService.countUserSubscribers(authorizedUser.getUserId()))
-                .addObject("friendsCount",userService.countUserFriends(authorizedUser.getUserId()));
+                .addObject("subscribersCount", userService.countUserSubscribers(authorizedUser.getUserId()))
+                .addObject("friendsCount", userService.countUserFriends(authorizedUser.getUserId()));
         UserDto user = searchService.getUserById(friendId);
 
         List<UserNews> userNews = userNewsService.getNews(friendId);
@@ -56,41 +56,49 @@ public class UserInfoController {
 
     @PostMapping("/home-page/search-user/user-info/{userId}")
     public ModelAndView userInfoDoView(@PathVariable("userId") String friendId,
-                                       @RequestParam("addFriendStatus") String status,
+                                       @RequestParam(name = "addFriendStatus", required = false) String status,
+                                       @RequestParam(name = "deleteNews", required = false) String deleteNewsId,
                                        @SessionAttribute(name = "authorizedUser", required = false) AuthorizedUser authorizedUser) {
 
 
         UserDto user = searchService.getUserById(friendId);
         ModelAndView modelAndView = new ModelAndView("user-info")
                 .addObject("user", user)
-                .addObject("subscribersCount",userService.countUserSubscribers(authorizedUser.getUserId()))
-                .addObject("friendsCount",userService.countUserFriends(authorizedUser.getUserId()));
+                .addObject("subscribersCount", userService.countUserSubscribers(authorizedUser.getUserId()))
+                .addObject("friendsCount", userService.countUserFriends(authorizedUser.getUserId()));
 
+        if (deleteNewsId != null) {
+            String oldStatus = friendsService.getFriendStatus(authorizedUser.getUserId(), friendId);
+            modelAndView.addObject("addFriendStatus", oldStatus);
+            userNewsService.deleteNews(deleteNewsId);
+        }
         List<UserNews> userNews = userNewsService.getNews(friendId);
+
         if (userNews != null) {
             modelAndView.addObject("userNews", userNews);
         }
 
-        if (authorizedUser.getUserRole().equals("ROLE_ADMIN") && status.equals("not_status")) {
-            friendsService.addFriend(authorizedUser.getUserId(), friendId, "friend");
-            return modelAndView.addObject("addFriendStatus", "friend");
-        }
+        if (status != null) {
+            if (authorizedUser.getUserRole().equals("ROLE_ADMIN") && status.equals("not_status")) {
+                friendsService.addFriend(authorizedUser.getUserId(), friendId, "friend");
+                return modelAndView.addObject("addFriendStatus", "friend");
+            }
 
-        if (status.equals("not_status")) {
-            friendsService.addFriend(authorizedUser.getUserId(), friendId, "subscriber");
-            return modelAndView.addObject("addFriendStatus", "subscriber");
-        }
+            if (status.equals("not_status")) {
+                friendsService.addFriend(authorizedUser.getUserId(), friendId, "subscriber");
+                return modelAndView.addObject("addFriendStatus", "subscriber");
+            }
 
-        if (status.equals("subscriber")) {
-            friendsService.deleteFriends(authorizedUser.getUserId(), friendId);
-            return modelAndView.addObject("addFriendStatus", "not_status");
-        }
+            if (status.equals("subscriber")) {
+                friendsService.deleteFriends(authorizedUser.getUserId(), friendId);
+                return modelAndView.addObject("addFriendStatus", "not_status");
+            }
 
-        if (status.equals("friend")) {
-            friendsService.deleteFriends(authorizedUser.getUserId(), friendId);
-            return modelAndView.addObject("addFriendStatus", "not_status");
+            if (status.equals("friend")) {
+                friendsService.deleteFriends(authorizedUser.getUserId(), friendId);
+                return modelAndView.addObject("addFriendStatus", "not_status");
+            }
         }
-
         return modelAndView.addObject("user", user);
     }
 }
